Rename dialect-agnostic coral-common classes from Hive to Coral
Create new Coral-named classes containing all logic for classes that
are not actually Hive-specific:
- HiveTypeSystem -> CoralTypeSystem
- HiveRelBuilder -> CoralRelBuilder
- HiveUncollect -> CoralUncollect
- LocalMetastoreHiveDbSchema -> LocalMetastoreDbSchema
- LocalMetastoreHiveSchema -> LocalMetastoreSchema
- RelDataTypeToHiveTypeStringConverter -> RelDataTypeToTypeStringConverter

Update ToRelConverter to use CoralTypeSystem and CoralRelBuilder.

Original Hive-named classes are retained as @Deprecated thin wrappers
for backward compatibility.
