+++ coral-common/src/main/java/com/linkedin/coral/common/LocalMetastoreDbSchema.java
@@ -1,0 +1,104 @@
+/**
+ * Copyright 2017-2026 LinkedIn Corporation. All rights reserved.
+ * Licensed under the BSD-2 Clause license.
+ * See LICENSE in the project root for license information.
+ */
+package com.linkedin.coral.common;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
+
+import org.apache.calcite.linq4j.tree.Expression;
+import org.apache.calcite.rel.type.RelProtoDataType;
+import org.apache.calcite.schema.*;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+
+
+/**
+ * Adaptor from a local metastore catalog providing database and table names
+ * to Calcite {@link Schema}.
+ *
+ * <p>This class is a replacement for {@link HiveDbSchema} to work with localMetastore in coral-spark-plan module.
+ */
+public class LocalMetastoreDbSchema implements Schema {
+
+  private final Map<String, Map<String, List<String>>> localMetastore;
+  private final String dbName;
+
+  public LocalMetastoreDbSchema(Map<String, Map<String, List<String>>> localMetastore, String dbName) {
+    checkNotNull(localMetastore);
+    checkNotNull(dbName);
+    this.localMetastore = localMetastore;
+    this.dbName = dbName;
+  }
+
+  @Override
+  public Table getTable(String tableName) {
+    if (localMetastore.containsKey(dbName) && localMetastore.get(dbName).containsKey(tableName)) {
+      return new LocalMetastoreHiveTable(tableName, localMetastore.get(dbName).get(tableName));
+    }
+    return null;
+  }
+
+  @Override
+  public Set<String> getTableNames() {
+    return ImmutableSet.copyOf(localMetastore.get(dbName).keySet());
+  }
+
+  @Override
+  public RelProtoDataType getType(String s) {
+    return null;
+  }
+
+  @Override
+  public Set<String> getTypeNames() {
+    return null;
+  }
+
+  @Override
+  public Collection<Function> getFunctions(String name) {
+    return ImmutableList.of();
+  }
+
+  @Override
+  public Set<String> getFunctionNames() {
+    return ImmutableSet.of();
+  }
+
+  /**
+   * A DB does not have subschema
+   * @param name name of the schema
+   * @return Calcite schema
+   */
+  @Override
+  public Schema getSubSchema(String name) {
+    return null;
+  }
+
+  @Override
+  public Set<String> getSubSchemaNames() {
+    return ImmutableSet.of();
+  }
+
+  @Override
+  public Expression getExpression(SchemaPlus parentSchema, String name) {
+    return null;
+  }
+
+  @Override
+  public boolean isMutable() {
+    return true;
+  }
+
+  // TODO: This needs to be snapshot of current state of catalog
+  @Override
+  public Schema snapshot(SchemaVersion schemaVersion) {
+    return this;
+  }
+}
